Further splitting up functionality
Maybe the names won't exactly stick, but I think it's useful to split
these up to avoid confusion. base is related to the individual emulator,
while interface is related to the actual use of emulator(s)
